% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/assignSubplots.R
\name{assignSubplots}
\alias{assignSubplots}
\title{Assigns the Correct Subplot Identity to Subplots}
\usage{
assignSubplots(df, subplotdf)
}
\arguments{
\item{df}{This is the major trees data.frame. It holds individual tree records, like in expandedTrees.}

\item{subplotdf}{This is the data.frame generated by buildBoxes, it has POINT_X, POINT_Y, and Subplot; Subplot will be overwritten by this function.}
}
\value{
This function returns a copy of subplotdf, with new values in Subplot.
}
\description{
This is a helper function to clean up the result from buildBoxes. It'll be wrapped up eventually.
}

