% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/buildBoxes.R
\name{buildBoxes}
\alias{buildBoxes}
\title{Build Boxes In A Plot Space}
\usage{
buildBoxes(x, y = NULL)
}
\arguments{
\item{x}{A vector of x values representing the corners of boxes along the
X-axis. If the x and y-values are identical, only include the first corner
vector, and the function will use it for the y-values as well.}

\item{y}{A vector of y values representing the corners of boxes along the
  Y-axis. See above for more detail about symmetrical plots.}
}
\value{
The parameters should be generated by getSubplotCoords, or you can
  enter them manually. This function returns a data.frame listing (x,y) of
  the corners of each subplot, arbitrarily numbered. The columns are
  "POINT_X", "POINT_Y", and "Subplot", reflecting the standard format taken
  by TrapUTM.
}
\description{
This function builds boxes based on corner vectors given by x and y.
}
\examples{
xAndY <- getSubplotCoords()
head(buildBoxes(xAndY, xAndY))
}

