% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_exposure.R
\name{calculate_exposure}
\alias{calculate_exposure}
\title{create a set of directories to run disperseR}
\usage{
calculate_exposure(
  year.E,
  year.D,
  link.to = "zips",
  pollutant = "SO2..tons.",
  units.mo,
  rda_file = "loaded",
  exp_dir = NULL,
  source.agg = c("total", "facility", "unit"),
  time.agg = c("year", "month"),
  return.monthly.data = F
)
}
\arguments{
\item{year.E}{emissions year}

\item{year.D}{HYSPLIT year}

\item{link.to}{spatial scale for plotting. One of 'zips', 'counties', or 'grids' that should match original input to disperseR::link_all_units()}

\item{pollutant}{name of column in `units.mo` for weighting}

\item{rda_file}{specification of data location produced by `disperseR::combine_monthly_links()`. Can denote 'loaded' if the object is loaded in the .GlobalEnv() or a file location}

\item{source.agg}{source aggregation level}

\item{time.agg}{time aggregation level}

\item{return.monthly.data}{logical. If TRUE, returns monthly datasets if `time.agg = 'month'`. Set to FALSE to save memory space.}

\item{exo_dir}{file path to save resulting exposure dataset}
}
\value{
Creates directories (does not overwrite if existing). Outputs string variables with paths to the environment.
}
\description{
`calculate_exposure()` takes linked HYSPLIT outputs combined using `disperseR::combine_monthly_links()` and sums them by emissions. The results can be aggregated to three different source levels and two different time scales. Various combinations of `year.E` and `year.D` can be applied to develop counterfactual scenarios.
}
\details{
\code{calculate_exposure}
}
