% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cart2polar.R
\name{cart2polar}
\alias{cart2polar}
\title{Convert Cartesian To Polar}
\usage{
cart2polar(x1, y1, x2, y2)
}
\arguments{
\item{x1}{The x-coordinate of your origin.}

\item{y1}{The y-coordinate of your origin.}

\item{x2}{A number or vector of numbers of destination points.}

\item{y2}{A number of vector of numbers of destination points.}
}
\value{
This function returns a data.frame with 4 columns. "x" is the
  distance between the xcoordinate and the origin, "y" is the distance
  between the y coordinate and the origin, theta is the angle from the
  original point, and r is the distance between the two points.
}
\description{
This is a function that converts a set of cartesian coordinates to a set of
polar coordinates.
}
\examples{
cart2polar(1,1,c(-2,2,1),c(11,1,2))
}

