% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/findRotation.R
\name{findRotation}
\alias{findRotation}
\title{Find Rotation Factor For Plot}
\usage{
findRotation(xleft, yleft, xbot, ybot)
}
\arguments{
\item{xleft}{The x value of the leftmost point.}

\item{yleft}{The y value of the leftmost point.}

\item{xbot}{The x value of the bottommost point.}

\item{ybot}{The y value of the bottommost point.}
}
\value{
This function returns a single numeric value representing the
  conversion factor needed to straighten a plot based on its leftmost and
  bottommost points. In essence, if the leftmost point is above the
  bottommost point in space, it will rotate counter clockwise until the y's
  of each point are equal. If they are already equal, no conversion will
  occur.
}
\description{
Given (x,y) of the left-most and bottom-most points in a dataframe, find the
angle required to rotate the plot to be straight. This function will return
the degrees of rotation required to turn a plot straight. Specifically, it
converts these two points into polar coordinates, then subtracts the angle of
their interaction from 360 degrees. Used primarily within other functions
like rotatePlot.
}
\examples{
findRotation(0,0,1,-1)
findRotation(0,2,0,0)
}

