% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/findSeedPlots.R
\name{findSeedPlots}
\alias{findSeedPlots}
\title{Finding All Possible Seedling Plots In An Area}
\usage{
findSeedPlots(seedlings, m)
}
\arguments{
\item{seedlings}{This is a data.frame with "x", "y", and "treeid" columns,
representing individual seedlings' geographic location and a unique
identifier.}

\item{m}{This is the search radius to search around each seedling to find its
  neighbors. The search parameter is a square, not a circle.}
}
\value{
This function returns a response data.frame with columns "x" and "y"
  to represent geographic location of the center of the box, and "n" with a
  number of seedlings found in that box.
}
\description{
This function searches a dataframe of seedlings and returns a list of (x,y)
coordinates where they are clumped together. The individuals entered into
this function must be of the same species, and the data.frame must contain an
"x" column, a "y" column, and a "treeid" column. The former two must be
numeric or integers.
}
\note{
This function follows these steps: Starting with the first row, it
  determines an m-bounded box around the (x,y) coordinates given. All
  individuals in the box (via a subset of the seedlings data.frame) are
  counted, and then removed from future searches to prevent double-counts.
  This algorithm may find slightly different results depending on the order
  of the seedlings in your data.frame. Plan accordingly. This function does
  not minimize the number of boxes drawn.
}
\examples{
## make a sample seed data.frame by subsetting the included
## expandedTrees data.frame..
sampleseeds <- expandedTrees[expandedTrees$plot=="trinity" &
                             expandedTrees$measyear==2001 &
                             expandedTrees$stage=="seedling",]
## show the start of results
head(findSeedPlots(sampleseeds, 1))
}

