% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/generatePlotMap.R
\name{generatePlotMap}
\alias{generatePlotMap}
\title{Generate Plot Map}
\usage{
generatePlotMap(species = 5, plotlength = 1000, plotwidth = 1000,
  maxsize = 100, propSeeds = 0.5, maxnum = 150)
}
\arguments{
\item{species}{The number of species to generate within the plot}

\item{plotlength}{The length of the plot (x axis)}

\item{plotwidth}{The width of the plot (y axis)}

\item{maxsize}{A number or vector of the max DBH size of the trees}

\item{propSeeds}{The number (or vector for each species) of the proportion of
trees generated that should be seedlings}

\item{maxnum}{A number of vector of number of trees per species}
}
\description{
This generates a random plot map for testing.
}
\examples{
generatePlotMap(species=3, plotlength=100, plotwidth=100,
maxsize=100, propSeeds=0.5, maxnum=20)
}

