% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getBoxBoundaries.R
\name{getBoxBoundaries}
\alias{getBoxBoundaries}
\title{Draw Box Around Point}
\usage{
getBoxBoundaries(x, y, buffer)
}
\arguments{
\item{x}{The x-coordinate of the center of your box.}

\item{y}{The y-coordinate of the center of your box.}

\item{buffer}{The distance, in coordinate space, of the searchradius}
}
\description{
This function returns four points around a search radius of an original x,y set.
}
\examples{
getBoxBoundaries(2, 5, 1)
}

