% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getBufferedPoints.R
\name{getBufferedPoints}
\alias{getBufferedPoints}
\title{Gets Points From A Rectangle With A Buffer}
\usage{
getBufferedPoints(x, y, buffer, n = 100, systematic = FALSE, by = 1)
}
\arguments{
\item{x}{A vector of x values}

\item{y}{A vector of y values}

\item{buffer}{A value that indicates how much buffer there should be along
each edge of the plot}

\item{n}{Optional. The number of points from the buffered area to sample from}

\item{systematic}{Optional, set to FALSE. If TRUE, this function will
systematically sample based on the value in the parameter "by"}

\item{by}{Optional. This tells the function how to increment during systematic
 sampling.}
}
\value{
This function returns a data.frame of x and y values within the buffer
 zone designated in the function call.
}
\description{
This function finds the min and max values of x and y in a dataframe, then
picks n number of points within that buffer and returns them in vector form.
}
\examples{
getRandomBufferedPoints(generatePlotMap()[,"x"],generatePlotMap()[,"y"],
buffer=10, n=5)
}

