% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getSubplotCoords.R
\name{getSubplotCoords}
\alias{getSubplotCoords}
\title{Get Points Within Min And Max Values By An Increment}
\usage{
getSubplotCoords(minmax = c(0, 100), increment = 25)
}
\arguments{
\item{minmax}{A vector of two points, representing the minimum and maximum values.}

\item{increment}{The increment value.}
}
\value{
This function returns a vector of numbers encompassing the min and max values provided.
}
\description{
This function takes "minmax", which is a vector with a minimum and maximum
value, and "increment", which defaults to 25, and goes up by that increment
until it goes just over the maximum value. This is a helper function to draw
boxes around points.
}
\examples{
getSubplotCoords()
}

