% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_all_units_zip.R
\name{link_all_units_zip}
\alias{link_all_units_zip}
\title{create a set of directories to run disperseR}
\usage{
link_all_units_zip(units.run, mc.cores = detectCores(),
  year.mons = NULL, start.date = NULL, end.date = NULL,
  pbl.height = pblheight, crosswalk. = crosswalk,
  duration.run.hours = 240, overwrite = FALSE)
}
\arguments{
\item{mc.cores}{`link_all_units_zip()` enables the parallel run by default splitting different months on different cores, but you can make it serial just by setting the `mc.cores` argument to `1`.As mentioned before  `link_all_units_zip()` enables the parallel run by default splitting different months on different cores, but you can make it serial just by setting the `mc.cores` argument to `1`.}

\item{year.mons}{these are the months for which we would like to do the linking. You can use the get_yearmon() function to create a vector that can be an input here.}

\item{start.date}{this argument is not necessary, but can be used if the user is interested in specifying a specific date to start the analysis with as opposed to using months. For example `start.date="2005-01-02"` for 2 January 2005. This argument are set to `NULL` by default and the function computes the start and the end dates using the `year.mons` provided.}

\item{pbl.height}{`pbl.height = pblheight` by default but you can change it. See the vignette}

\item{crosswalk.}{`crosswalk. = crosswalk` by default but you can change it. See the vignette}

\item{start.end}{this argument is not necessary, but can be used if the user is interested in specifying a specific date to end the analysis with as opposed to using months. For example `start.date="2005-01-02"` for 2 January 2005.This argument are set to `NULL` by default and the function computes the start and the end dates using the `year.mons` provided.}
}
\value{
vector of months that you can loop over
}
\description{
with `link_all_units_zip()` users can link all air parcels to ZIP codes by month for specified units. with combinations of years and months. `link_all_units_zip()` uses another function called `disperser_link_zip()` that runs in the background and reads in all the relevant files (i.e., those that correspond to the provided units) produced by the `run_disperser_parallel()` function and saves them. Then it links them to ZIP codes.
}
\details{
\code{link_all_units_zip}
}
