% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impact_unit.R
\name{plot_impact_unit}
\alias{plot_impact_unit}
\title{create a set of directories to run disperseR}
\usage{
plot_impact_unit(
  data.linked = NULL,
  zip.codes = NULL,
  y.lim = c(24, 50),
  x.lim = c(-123, -69),
  plot.title = NULL,
  legend.lims = NULL,
  legend.title = NULL,
  legend.text.angle = 0,
  graph.dir = NULL
)
}
\arguments{
\item{data.linked}{data for plotting as output from disperseR::calculate_exposure()}

\item{zip.codes}{list of ZIP codes (as character strings) for ranking}

\item{y.lim}{vertical limits for the spatial plot}

\item{x.lim}{horizontal limits for the spatial plot}

\item{plot.title}{plot title character string}

\item{legend.lims}{legend limits. as c( lower,upper)}

\item{legend.title}{legend title as string}

\item{legend.text.angle}{angle of legend text (helpful for large numbers)}

\item{graph.dir}{location to save output.}
}
\value{
Creates directories (does not overwrite if existing). Outputs string variables with paths to the environment.
}
\description{
`plot_impact_unit()` produces two graphs. One is a map showing where the units are and showing their rank and the other one contains bar plots.
}
\details{
\code{plot_impact_unit}
}
