% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impact_weighted.R
\name{plot_impact_weighted}
\alias{plot_impact_weighted}
\title{\code{plot_impact_weighted}}
\usage{
plot_impact_weighted(
  data.linked,
  data.units,
  link.to = "zips",
  zcta.dataset = NULL,
  counties. = NULL,
  metric = "hyads",
  time.agg = "year",
  legend.lims = NULL,
  legend.name = NULL,
  plot.name = NULL,
  map.month = NULL,
  graph.dir = NULL,
  zoom = TRUE,
  ...
)
}
\arguments{
\item{data.linked}{data for plotting as output from disperseR::calculate_exposure()}

\item{data.units}{information on unit locations as output from disperseR::units()}

\item{link.to}{spatial scale for plotting. One of 'zips', 'counties', or 'grids' that should match original input to disperseR::calculate_exposure()}

\item{zcta.dataset}{ZIP code/ZCTA sptial dataset as imported by disperseR::get_data() or subset thereof. Required only if link.to = 'zips'.}

\item{counties.}{US counties dataset as imported using USAboundaries::us_counties( ) or subset thereof. Required only if link.to = 'counties'.}

\item{metric}{plotting metric in `data.linked` input. Defaults to 'hyads', the default output of disperseR::calculate_exposure().}

\item{time.agg}{time aggregation matching original input to disperseR::calculate_exposure(). Can be 'month' or 'year'}

\item{legend.lims}{legend limits string. Defaults to NULL, which creates a scale from 0 to the 95th percentile of the 'metric'. Should be in the form c(lower,upper).}

\item{legend.name}{legend name string. Defaults to NULL, which creates a legend title of 'Aggregate HyADS Exposure'.}

\item{plot.name}{plot title string. Defaults to NULL, or a blank title}

\item{map.month}{specific month to map in form YYYYMM if time.agg = 'month'. Can be created using disperseR::get_yearmon()}

\item{graph.dir}{location to save output.}

\item{...}{modeling parameters passed to ggplot2::theme()}
}
\value{
Creates ggplot object of spatial HyADS impacts
}
\description{
`plot_impact_weighted()` takes output from `calculate_exposure()` and creates a spatial plot
}
