% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/polar2cart.R
\name{polar2cart}
\alias{polar2cart}
\title{Convert Polar Coordinates To Cartesian Coordinates}
\usage{
polar2cart(x, y, dist, bearing, as.deg = TRUE, rnd = NA)
}
\arguments{
\item{x}{The origin x-coordinate. Numeric.}

\item{y}{The origin y-coordinate. Numeric.}

\item{dist}{A number or vector of numbers of distances from the origin.}

\item{bearing}{A number or vector of numbers representing the degrees away
from the origin. This follows traditional polar coordinates, with 0 at
"east", 90 at "north", etc.}

\item{as.deg}{Default set to true, if false, then the program assumes that
"bearing" is entered in radians instead of degrees.}

\item{rnd}{The rounding value. Default is NA, for no rounding. Set to 0 for integers, 1 for 1 decimal point, etc.}
}
\value{
This function returns a list with "x" and "y" coordinates for the polar coordinates entered.
}
\description{
This function converts polar coordinates to cartesian coordinates. It needs
an origin (x,y), as well as an individual number or vector of numbers for
distance and bearing.
}
\examples{
polar2cart(1, 1, 2, 90)
polar2cart(1,1,c(2,2,2),c(90,180,270) )
}

