% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_dispersion_subfun.R
\name{read_listing_file}
\alias{read_listing_file}
\title{Read a `listing` file}
\usage{
read_listing_file(file_path)
}
\arguments{
\item{file_path}{The path to the \code{listing} file.}
}
\description{
Read a `listing` file
}
