% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_disperser_parallel.R
\name{run_disperser_parallel}
\alias{run_disperser_parallel}
\title{run the fac model in parallel}
\usage{
run_disperser_parallel(
  input.refs = NULL,
  pbl.height = NULL,
  species = "so2",
  proc_dir = proc_dir,
  overwrite = F,
  npart = 100,
  mc.cores = parallel::detectCores(),
  keep.hysplit.files = FALSE
)
}
\arguments{
\item{input.refs}{should be the data table that is the result of the `define_input()` function or a subset of that dataset}

\item{pbl.height}{Monthly mean planetary boundary layer heights. See vignettes for more information}

\item{species}{The package has the possibility to use two types of species. The default one is `species = 'so2'`, but you can also use particulate sulfate `species = 'so4p'`.}

\item{proc_dir}{directory where the function saves temporary files while running. This is automatically defined by `create_dirs()`}

\item{overwrite}{if output files already exist should they be overwritten? This is `false` by default.}

\item{npart}{number of air parcels tracked by HYSPLIT. Defaults to 100}

\item{mc.cores}{on how many cores should R split the computations. set to  parallel::detectCores() or set to 1 if you want to serial computation.}

\item{keep.hysplit.files}{logical. If FALSE (the default), clears storage space in the `proc_dir` by removing all HYSPLIT files}
}
\value{
This function returns fac model results.
}
\description{
\code{run_disperser_parallel}
}
\details{
It is possible that running the below code with output a warning "WARNING: map background file not found ../graphics/arlmap". It is safe to ignore it.
}
