% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/sampleSubplots.R
\name{sampleSubplots}
\alias{sampleSubplots}
\title{Sample Seedling Subplots For Seedlings}
\usage{
sampleSubplots(subplotxy, fulldf, subplotsize = 1)
}
\arguments{
\item{subplotxy}{A data.frame of x and y coordinates for sampling subplots}

\item{fulldf}{The full plot dataframe, containing all trees and seedlings
within your larger plot space.}

\item{subplotsize}{The size of the subplot in coordinate space, default is 1.}
}
\description{
This function takes a list of random x and y coordinates within a plot,
builds a box around each point, and subsets a full plot's dataframe to see if
there are any seedlings inside.
}

