% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ssdPlotDesc.R
\docType{data}
\name{ssdPlotDesc}
\alias{ssdPlotDesc}
\title{ssdPlotDesc Dataset}
\format{A data frame with 14 rows and 12 variables: \describe{
  \item{plot}{The plot's nickname.}
  \item{elev}{The elevation of the plot.}
  \item{size_ha}{The size of the plot, in hectares.}
  \item{established}{The year that the plot was established.}
  \item{yr1}{The first measurement year.}
  \item{yr2}{The second measurement year.}
  \item{yr3}{The third measurement year.}
  \item{yr4}{The fourth measurement year.}
  \item{yr5}{The fifth measurement year.}
  \item{yr6}{The sixth measurement year.}
  \item{yr7}{The seventh measurement year.}
  \item{range}{The length of time that this plot has been measured.}
  }}
\usage{
ssdPlotDesc
}
\description{
This data.frame describes the plots found in ssdAllTrees. These data (in both
sets) are part of a larger dataset that was obfuscated for privacy.
}
\keyword{datasets}

