% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/trapUTM.R
\name{trapUTM}
\alias{trapUTM}
\title{Find SeedTrap Locations}
\usage{
trapUTM(filename, subplots, site, bearing = 0, plotcorns = list(F))
}
\description{
This function is imported from other authors. It will be updated with function soon enough.
}

