% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{create a set of directories to run disperseR}
\usage{
create_dirs(location = file.path("~", "Desktop"))
}
\arguments{
\item{location}{this is the path to where you want your set of directories to be created. By default this is set to desktop.}
}
\value{
Creates directories (does not overwrite if existing). Outputs string variables with paths to the environment.
}
\description{
`create_dirs()` takes 0 or the following argument `location`
}
\details{
\code{create_dirs}
}
